update resume link
now we link to the raw resume pdf in the cv repo (github.com/michsa/cv).
this shouldn't need to be updated again in a very long time.

also updates the descrip a little bit. i've been a SENIOR SOFTWARE
ENGINEER for a while now. finally retiring "wizard errant" because
ddg'ing it just turns up a bunch of harry potter fanfiction (it's
actually a reference to diane duane's "young wizards" series fyi).